refactor: move ResponseEntityUtil to server.annotation package
